+++ kogito-quarkus-examples/process-usertasks-timer-quarkus-with-console/docker-compose/startServices.sh
@@ -23,12 +23,12 @@ elif [ "$(expr substr $(uname -s) 1 5)" == "Linux" ]; then
    echo "DOCKER_GATEWAY_HOST=172.17.0.1" >> ".env"
 fi
 
-BD="postgresql"
+DB="postgresql"
 
 if [ -n "$1" ]; then
-  if [[ "$1" == "postgresql" || "$1" == "mongodb" || "$1" == "infinispan" || "$1" == "p" ]];
+  if [[ "$1" == "postgresql" || "$1" == "infinispan" ]];
   then
-    BD="$1"
+    DB="$1"
   else
    echo "Usage: By default postgresql environments is started if no argument is provided"
    echo "     start POSTGRESQL docker-compose running: ./startServices.sh postgresql or just ./startServices.sh "
@@ -36,22 +36,24 @@ if [ -n "$1" ]; then
    exit 1
   fi
 fi
-echo "Have you compiled the project before with the right profile: ../mvn clean install -DskipTests -P$BD"
+echo "Have you compiled the project before with the right profile: ../mvn clean install -DskipTests -P$DB"
 
-PERSISTENCE_FOLDER=./persistence
-KOGITO_EXAMPLE_PERSISTENCE=../target/classes/META-INF/resources/persistence/protobuf
-
-rm -rf $PERSISTENCE_FOLDER
-
-mkdir -p $PERSISTENCE_FOLDER
-
-if [ -d "$KOGITO_EXAMPLE_PERSISTENCE" ]
+if [ "$1" == "infinispan" ];
 then
-    cp $KOGITO_EXAMPLE_PERSISTENCE/*.proto $PERSISTENCE_FOLDER/
-else
-    echo "$KOGITO_EXAMPLE_PERSISTENCE does not exist. Have you compiled the project? mvn clean install -DskipTests -P$BD"
+  PERSISTENCE_FOLDER=./persistence
+  KOGITO_EXAMPLE_PERSISTENCE=../target/classes/META-INF/resources/persistence/protobuf
 
+  rm -rf $PERSISTENCE_FOLDER
+
+  mkdir -p $PERSISTENCE_FOLDER
+
+  if [ -d "$KOGITO_EXAMPLE_PERSISTENCE" ]
+  then
+    cp $KOGITO_EXAMPLE_PERSISTENCE/*.proto $PERSISTENCE_FOLDER/
+  else
+    echo "$KOGITO_EXAMPLE_PERSISTENCE does not exist. Have you compiled the project? mvn clean install -DskipTests -P$DB"
     exit 1
+  fi
 fi
 
 SVG_FOLDER=./svg
@@ -68,4 +70,4 @@ else
     exit 1
 fi
 
-docker-compose -f docker-compose-$BD.yml up
+docker-compose -f docker-compose-$DB.yml up
